pyproject.toml
[project]
name = "ai-bestie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "generativeai>=0.0.1",
    "google-generativeai>=0.8.5",
    "openai[agents]>=1.108.1",
]
